Module 7 Check App Status
This project checks if the DesktopUI App is working to Withdrawal section.

If the Withdrawal section is checked then it writes an amount to withdrawal.

If isn't checked then display a MessageBox, change the section clicking in Withdrawal and write the amount to withdrawal.

Also, has a process to check if the file 'Wirhdrawal Details.txt" is open, displaying a Message Box in both cases.