Module 8 Practice02 Actions and Output Methods
Workflow that navigates to the movie search page of the RPA Challenge website and retrieves only the first column of the results page to display a full description of the movie (not seen on screen) and the movie image text along with description. that is seen on the screen. Then compare the outputs.