Module 8 Practice 3 Login and DataScraping Collect
Suppose we need to collect data to create a consolidated report to present to stakeholders. What do we do when our data is spread across multiple pages or, for example, across multiple pages on different websites? Manually extracting all this data will probably take a lot of time and effort. Your task is to automate a process that extracts the following data: WIID, ​​description, type, and status from the ACME Work Items page, using the data extraction method. Then, save the extracted data to an Excel file.