Module 10 Practice 1 Extract Email Without Regex
On a string containing an email address, create a workflow to extract the email address without using regular expressions.

Note: For data entry, use a string variable with the following value "Please use the following address to contact me john.doe@localcompany.com, this is the company email."

Steps:

Create a new process that reads and extracts the email address from the provided string.
Show the email address in a Register Message activity.