Module 10 Practice 2 Get Email With Regex
The human resources department is using a template to request the creation of new employee IDs. Every day the template is updated with new email addresses and sent for processing.

Use the provided template file to create a workflow that searches for email addresses within the template using a regular expression.

Note: As input data, use the provided template file, Template.txt.