Module 11 Practice 02 Dedine Data
Problem Statement: You have an Excel file with some data. Your task is to create a workflow that extracts information from the Excel file and populates all the information into the fields of a website. Manual processing consumes more time to complete the job. Therefore, it is a good idea to automate the entire process.