Module 12 Practice01FixMyWorkflow
You have received the following workflow from a colleague. The process is not running correctly and you must correct it.

Download the project and correct the workflow using the concepts and features you have studied in this course.

The workflow should do the following:

Search for the movie "Ice Age" on the IMDb website.
Get IMDb score.
Show it through a message box.
Expected output: Message box showing the score for the movie Ice Age.