Module 19 Practice 01 Outlook Mail Organization
Create a project that reads the inbox and sorts emails by moving them to various folders, based on "rules" defined in an Excel similar to the one shown below.

If the sender address contains the string defined in the Sender column, move the mail to the email folder defined in the Folder column.